refactor philo
