+++ philo/inc/actors/philo_actor.h
@@ -6,7 +6,7 @@
 /*   By: ttsubo <ttsubo@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/06/06 16:28:12 by ttsubo            #+#    #+#             */
-/*   Updated: 2025/06/20 22:22:25 by ttsubo           ###   ########.fr       */
+/*   Updated: 2025/06/21 18:34:50 by ttsubo           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -50,6 +50,9 @@ typedef struct s_philo_actor
 
 t_philo_actor	*philo_actor_new(size_t no, t_main_args args);
 void			philo_actor_del(t_philo_actor **philo_ptr);
+void			philo_on_start(t_ft_actor *self);
+bool			philo_on_receive(t_ft_actor *self, t_ft_msg *msg);
+// methods
 void			_common_update(t_philo_actor *self, t_ft_msg *msg);
 void			_thinking(t_philo_actor *self);
 void			_eating(t_philo_actor *self);
