+++ philo/srcs/actors/philo/_sleeping.c
@@ -1,0 +1,23 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   _philo_actor_methods1.c                            :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: ttsubo <ttsubo@student.42.fr>              +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2025/06/06 16:32:30 by ttsubo            #+#    #+#             */
+/*   Updated: 2025/06/21 18:37:32 by ttsubo           ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
+#include "philo_actor.h"
+
+void	_sleeping(t_philo_actor *self)
+{
+	self->now_slp++;
+	if (self->now_slp >= self->max_slp)
+	{
+		self->now_slp = 0;
+		self->sts = PHILO_STS_THINKING;
+	}
+}
