+++ philo/srcs/actors/philo/philo_actor.c
@@ -6,58 +6,12 @@
 /*   By: ttsubo <ttsubo@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/06/06 16:32:30 by ttsubo            #+#    #+#             */
-/*   Updated: 2025/06/20 22:22:38 by ttsubo           ###   ########.fr       */
+/*   Updated: 2025/06/21 18:35:06 by ttsubo           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "main.h"
 #include "philo_actor.h"
-
-static void	on_start(t_ft_actor *self)
-{
-	t_philo_actor	*philo;
-	t_ft_actor		*parent;
-
-	philo = self->ref;
-	parent = self->parent;
-	philo->sts = PHILO_STS_THINKING;
-	philo->now_hp = philo->max_hp;
-	philo->now_eat = 0;
-	philo->now_slp = 0;
-	philo->eat_count = 0;
-	philo->has_r_fork = false;
-	philo->has_l_fork = false;
-	philo->can_eat = false;
-	parent->tell(parent, msg_new(ACTOR_START_DONE, self, NULL));
-}
-
-static bool	on_receive(t_ft_actor *self, t_ft_msg *msg)
-{
-	t_philo_actor	*philo;
-
-	if (!self || !self->ref || !msg)
-		return (true);
-	philo = (t_philo_actor *)self->ref;
-	if (msg->type == UPDATE)
-	{
-		if (philo->sts == PHILO_STS_DEAD)
-			return (true);
-		_common_update(philo, msg);
-		if (philo->sts == PHILO_STS_THINKING)
-			_thinking(philo);
-		if (philo->sts == PHILO_STS_EATING)
-			_eating(philo);
-		if (philo->sts == PHILO_STS_SLEEPING)
-			_sleeping(philo);
-	}
-	if (msg->type == GRANT_EAT)
-		philo->can_eat = true;
-	if (msg->type == GRANT_FORK)
-		_confirm_fork(philo, msg->sender);
-	if (msg->type == FORK_RELEASED)
-		_on_fork_released(philo, msg->sender);
-	return (true);
-}
 
 void	philo_actor_del(t_philo_actor **philo_ptr)
 {
@@ -75,8 +29,8 @@ t_philo_actor	*philo_actor_new(size_t no, t_main_args args)
 {
 	t_philo_actor					*philo;
 	static const t_ft_actor_vtable	vtable = {
-		.on_start = on_start,
-		.on_receive = on_receive,
+		.on_start = philo_on_start,
+		.on_receive = philo_on_receive,
 		.on_stop = NULL,
 	};
